dataPipeline-main/pipeline/requirements.txt
byteplus-python-sdk-v2
python-dotenv
requests
opencv-python-headless
